// project/web/src/graphql/api/operations.ts
export type FilmsQuery = {
  __typename?: 'Query';
  films: {
    __typename?: 'PaginatedFilms';
    cursor?: number | null;
    films: Array<{
      __typename?: 'Film';
      id: number;
      title: string;
      subtitle?: string | null;
      runningTime: number;
      releaseDate: string;
      posterImg: string;
      director: { __typename?: 'Director'; name: string };
    }>;
  };
};